// src/app/models/patient.model.ts
export class Patient{
    ID: number;
    PassNumber: string;
    ForeNames: string;
    SurName : string;
    DateOfBirth: Date; 
    Gender: string;
    HomeTelephoneNumber: string;
    GeneralPartitioner:string;
    NOKName: string;
    relationShip: string;
    NOKAddress1: string;
    NOKAddress2: string;
    NOKAddress3: string;
    NOKAddress4: string;
    GPSurName:string;
    
}